Rename subject directory to model
