+++ core/src/model/util.ts
